refactor : 이제 요청들을 각자 handler들을 가지고 처리할 수 있도록 수정
HTTP 메서드들과 리소스 요청들을 처리하는 핸들러들을 만들었기 때문에, 응답을 만들기 위해 해당 핸들러들을 이용해 처리하도록 리펙토링
